refactor(memory-core): separate dreaming state codec ownership
